requirements.txt
numpy>=1.18.0
torch>=1.8.0
nibabel
scipy
matplotlib
tqdm
scikit-image
pyyaml
Pillow